pyproject.toml
[tool.poetry]
name = "discord_itk_bot"
version = "0.1.0"
description = "Just a stupid discord bot for my personal use."
authors = ["Kerati <enderwolf50@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.6"
"discord.py" = "^1.7.3"
pymongo = "^3.12.0"
sentry-sdk = "^1.3.0"
coloredlogs = "^15.0.1"
PyYAML = "^5.4.1"
saucenao-api = "^2.3.1"
google-api-python-client = "^2.13.0"
addict = "^2.4.0"
dnspython = "^2.1.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
